Refactor Supabase client imports and update project dependencies
