+++ app/actions.ts
@@ -1,6 +1,6 @@
 "use server"
 
-import { createClient } from "@/lib/supabase"
+import { createClient } from "@/lib/supabase-server"
 import { revalidatePath } from "next/cache"
 import type { Note, Notebook, Tag } from "@/lib/types"
