+++ lib/supabase-client.ts
@@ -1,0 +1,6 @@
+import { createBrowserClient } from "@supabase/ssr"
+
+// Client-side client for Client Components
+export function createClientComponentClient() {
+  return createBrowserClient(process.env.NEXT_PUBLIC_SUPABASE_URL!, process.env.NEXT_PUBLIC_SUPABASE_ANON_KEY!)
+}
