+++ lib/supabase-server.ts
@@ -1,5 +1,4 @@
 import { createServerClient } from "@supabase/ssr"
-import { createBrowserClient } from "@supabase/ssr"
 import { cookies } from "next/headers"
 
 // Server-side client for Server Components and Server Actions
@@ -22,9 +21,4 @@ export async function createClient() {
       },
     },
   })
-}
-
-// Client-side client for Client Components
-export function createClientComponentClient() {
-  return createBrowserClient(process.env.NEXT_PUBLIC_SUPABASE_URL!, process.env.NEXT_PUBLIC_SUPABASE_ANON_KEY!)
 }
